+++ src/libopenrave-core/colladaparser/colladawriter.cpp
@@ -1789,7 +1789,7 @@ private:
                 pcage->add("sidewallExtents")->setCharData(ss.str());
                 ss.clear(); ss.str("");
 
-                uint8_t b = info._sidewallExists;
+                uint32_t b = info._sidewallExists;
                 ss << b;
                 pcage->add("sidewallExists")->setCharData(ss.str());
